Refactor: Use f32 for pheromone deposition and decay
This commit changes the data type of `agentDepositionAmount` and `pheromoneDecayFactor` from `u8` to `f32`. This allows for more granular control over these simulation parameters.

The following files were updated:
- `simulation_settings.toml`: Adjusted default values.
- `src/agent.rs`: Updated `Agent` struct and related methods.
- `src/pheromones.rs`: Updated `Pheromones` struct and related methods, including scaling `f32` deposition to `u8` for grid storage.
- `src/settings.rs`: Updated `Settings` struct and default constant values.

+++ src/agent.rs
@@ -32,7 +32,7 @@ pub struct Agent {
     jitter: f32,
     #[builder(default = default_rng())]
     rng: StdRng,
-    deposition_amount: u8,
+    deposition_amount: f32,
 }
 
 #[derive(Default)]
@@ -44,7 +44,7 @@ pub struct AgentUpdate {
     pub move_speed: Option<f32>,
     pub rotation_speed: Option<f32>,
     pub jitter: Option<f32>,
-    pub deposition_amount: Option<u8>,
+    pub deposition_amount: Option<f32>,
 }
 
 impl Agent {
@@ -117,7 +117,7 @@ impl Agent {
         self.location
     }
 
-    pub fn deposition_amount(&self) -> u8 {
+    pub fn deposition_amount(&self) -> f32 {
         self.deposition_amount
     }
